opengl within defapplet works
Processing's internal exception was thrown when trying to set
the renderer to anything else than JAVA2D (looks like issuse with
order of initialization of different components).

Now the :size option ignores the third element of the vector (if supplied).
Use (size 200 200 OPENGL) within the setup method  to set size and renderer.

Tested with Processing 1.5.1