Add internal state function to core.cljc
Internal state is an atom added to every sketch. It's a handy way to store sketch-specific Quil state both in clj and cljs. Currently such pieces of state are stored on in meta object in Clojure or on processing object in cljs which leads to conditional macros in cljs an makes code harder to read. Internal state helps with. The downside that it's probably slower compared to setting properties on object but speed is not a big concern in the moment. We can optimize it later if needed.

This commit also migrates existing scattered states to the internal state.

Tested:
  manually