Add test helper with-sketch that dispatches between clj/cljs applet bindings
For reasons, quil.sketch/*applet* is the global active applet in cljs, and
quil.applet/*applet* for clj processing sketches.

Attempted this with reader conditionals and had lots of issues, so manually
check during macro compilation if it's a clj/cljs context evaluating using the
&env environment hack.

Without this fix, cljs tests would run all the internal assertions, but then in
an applet test, clj would ignore all the inner assertions of the body of
with-sketch. So if modifying this, pay close attention to the total number of
assertions reported. If this starts breaking, tests will still finish, but it
will have ignored inner assertions.