Fix clj snippet tests swallowing test failures
t/is check *must* occur on same thread as test reporting, not on the quil render
thread. So now it saves the image, delivers a result and then verifies it
matches the output.

One caveat, this meaning in manual mode the verification happens after the
sketch closes, so might make sense to run the verify step if manual is on.

Alternatively, figure out a better way to keep the image visible.