Pass through integer colors for background, color, fill and stroke
Partial fix for #364. Forces any of the mulity-arity color functions to preserve
integers on gray or gray,alpha argument signatures.

This is sufficient to fix 5 tests so it's a good stop gap, but not happy with
performance and I still think a table of values and expected would help for this.