increase tolerance but use reference images from github actions runner
The new expected values are what a runner on github calculated, instead of
weirdness locally. I suspect the tolerance will actually need to be set per
snippet but not doing that yet. The sphere/detail tests and the ambient-light
test are all an anti-aliasing issue but I can't figure out why.