Add a trailing q/no-loop to every snippet using create-{graphics,image}
For the CLJ snippet tests, we only run draw once before saving a snapshot.
However, for cljs, it async starts the sketch on browsing the page and then
loops forever. The etaoin snapshots may or may not happen on the first frame.
The side effect of this is a new canvas or image is created on every loop, so
this stops the loop after a single iteration.