Actually support adjusting accepted-diff-threshold for cljs
Currently re-querying the list of examples with etoain, so the extra params have
to be threaded through as data attributes on the options.

Long term this should do something like annotate the snippets list with the
option index for the test instead.

Also extracts snippet-elements helper out of the inner test loop