set background after with-graphics to fix compositing
Not clear why this is necessary but certain compositing actions with another
graphics instance on a white background are reverting to use the gray
background.

This is also updating the reference as it is now correctly showing the green
circle on the red square before clearing the image to composite with.