experiment updating to jogl/gluegen 2.5.0
Processing is still running an earlier version, however there are a number of
segv cases encountered in our test suite, and it *does* pass using the updated
dependencies, so trying this to see if it works.

This does not address the thread lock/hanging issue, but hopefully will assist
with segv.