try slowing down the framerate in CI to reduce SEGV
When running under github actions, reduce frame rate to 10fps. Previously the
majority of the errors were flakes at 5fps, but without that limit encountering
a lot of SEGV errors. So the experiment is to see if 100ms per frame is more
stable for SEGV issues in the renderers. Hypothesis is that there is a race
condition on cleaning up the graphics context and sometimes that triggers a
segfault.