late bind to figwheel so that we can load all tests
When running cljs tests the alias :fig is used to bring figwheel and various
other dependencies into the classpath. However we don't strictly need any of
those when running the default tests, so it's preferable to bind to figwheel
only at the time of test invocation, not at namespace load.